+++ src/App/SelectMultiple.fs
@@ -35,7 +35,7 @@ let rec join (flavours : string list) =
 
 // HTML helpers
 let block children =
-    Html.div <| (class' "block") :: children
+    Html.div ((class' "block") :: children )
 
 // Control with only 1 label child
 let controlLabel children =
@@ -59,7 +59,7 @@ let view() =
                     Html.input [
                         type' "radio"
                         Bindings.bindRadioGroup scoops
-                        i+1 |> string |> value
+                        i+1 |> string |> Attr.value
                     ]
                     text $" {scoopChoice}"
                 ]) |> fragment
@@ -70,12 +70,12 @@ let view() =
             Html.div [
                 class' "select is-multiple"
                 Html.select [
-                    multiple
+                    Attr.multiple true
                     Bindings.bindSelectMultiple flavours
                     fragment [
                         for flavour in menu do
                             Html.option [
-                                value flavour
+                                Attr.value flavour
                                 text $" {flavour}"
                             ]
                     ]