+++ src/App/StaticEachWithIndex.fs
@@ -3,9 +3,7 @@ module StaticEachWithIndex
 // Adapted from
 // https://svelte.dev/examples
 
-open Sutil
 open Sutil.DOM
-open Sutil.Attr
 
 type Cat = { Id : string; Name : string }
 
@@ -29,8 +27,8 @@ let view() =
             for (i,cat) in (withIndex cats) do
                 Html.li [
                     Html.a [
-                        target "_blank"
-                        href $"https://www.youtube.com/watch?v={cat.Id}"
+                        Attr.target "_blank"
+                        Attr.href $"https://www.youtube.com/watch?v={cat.Id}"
                         text $"{i + 1}: {cat.Name}"
                     ]
                 ]